make directory and move file for Sleep habbit, did some cleanup and fixed linting error
