+++ src/components/SleepHabbit/Timer.js
@@ -3,10 +3,7 @@ import { StyleSheet, View, Text } from 'react-native';
 import { connect } from 'react-redux';
 import PercentageCircle from 'react-native-percentage-circle';
 import { StackActions } from 'react-navigation';
-import { changeInterval, countdown } from '../actions';
-
-// If you want to check access to Local Storage, add function from helper and add Button component import from react-native.
-// import { retrieveEndTime } from '../helper';
+import { changeInterval, countdown } from '../../actions';
 
 const styles = StyleSheet.create({
   container: {
@@ -36,6 +33,7 @@ export class Timer extends Component {
 
   shouldComponentUpdate(nextProps) {
     const { goToYay } = this.props;
+    // eslint-disable-next-line
     if (nextProps.currentCounter._milliseconds === 0) {
       goToYay();
     }
@@ -52,32 +50,6 @@ export class Timer extends Component {
     const time = currentCounter.format('h:mm:ss');
     return (
       <View style={styles.container}>
-        {/* If you want to check access to LocalStorage Uncomment Button component */}
-        {/* <Button
-          onPress={sayHi}
-          title="Don't push me"
-          color="#fff"
-          accessibilityLabel="Don't do that"
-        />
-        <Button
-          onPress={setEndTimer}
-          title="Save data to LS"
-          color="#f6eaea"
-          accessibilityLabel="Don't do that"
-        />
-        <Button
-          onPress={retrieveEndTime}
-          title="Retrieve data from LS"
-          color="#f6eaea"
-          accessibilityLabel="Don't do that"
-        />
-        <Button
-          onPress={removeEndTime}
-          title="Bomberman"
-          color="#f6eaea"
-          accessibilityLabel="Don't do that"
-        />
-        <Text> {percentage} </Text>{' '} */}
         <PercentageCircle
           borderWidth={14}
           radius={100}
